organizing heap implementation into separate files
